fix: move low-risk fsd dependencies to shared
