+++ Testing/unit/features/tasks/components/TaskDetailsView.deleteGuard.test.tsx
@@ -24,7 +24,7 @@ vi.mock('@/features/tasks/hooks/useTaskComments', () => ({
 vi.mock('@/features/tasks/hooks/useTaskCommentsRealtime', () => ({
     useTaskCommentsRealtime: () => undefined,
 }));
-vi.mock('@/features/projects/hooks/useProjectActivity', () => ({
+vi.mock('@/shared/hooks/useActivityLog', () => ({
     useProjectActivity: () => ({ data: [], isLoading: false }),
     useTaskActivity: () => ({ data: [], isLoading: false }),
 }));
