+++ src/features/projects/hooks/useProjectPresence.ts
@@ -1,15 +1,7 @@
 import { useEffect, useRef, useState } from 'react';
 import { supabase } from '@/shared/db/client';
 import { useAuth } from '@/shared/contexts/AuthContext';
-
-export interface PresenceState {
-    user_id: string;
-    email: string;
-    /** ms epoch — used for stable sort and same-user dedup across tabs. */
-    joinedAt: number;
-    /** Peer's currently-focused task id, or null when viewing the project shell. */
-    focusedTaskId: string | null;
-}
+import type { PresenceState } from '@/shared/types/presence';
 
 interface UseProjectPresenceResult {
     presentUsers: PresenceState[];
