+++ src/features/tasks/components/TaskDetailsView.tsx
@@ -7,8 +7,8 @@ import * as z from 'zod';
 import TaskResources from '@/features/tasks/components/TaskResources';
 import TaskDependencies from '@/features/tasks/components/TaskDependencies';
 import TaskComments from '@/features/tasks/components/TaskComments/TaskComments';
-import { useTaskActivity } from '@/features/projects/hooks/useProjectActivity';
-import { ActivityRow } from '@/features/projects/components/ActivityRow';
+import { useTaskActivity } from '@/shared/hooks/useActivityLog';
+import { ActivityRow } from '@/shared/ui/ActivityRow';
 import { formatDisplayDate } from '@/shared/lib/date-engine';
 import { useAuth } from '@/shared/contexts/AuthContext';
 import { useTaskSiblings } from '@/features/tasks/hooks/useTaskSiblings';
@@ -28,7 +28,7 @@ import type { TaskItemData } from '@/features/tasks/components/TaskItem';
 import type { TaskRow, TeamMemberWithProfile } from '@/shared/db/app.types';
 import { extractCoachingFlag } from '@/features/tasks/lib/coaching-form';
 import { extractStrategyTemplateFlag } from '@/features/tasks/lib/strategy-form';
-import { extractPhaseLeads } from '@/features/projects/lib/phase-lead';
+import { extractPhaseLeads } from '@/shared/lib/phase-lead';
 import { useTeam } from '@/features/people/hooks/useTeam';
 import StrategyFollowUpDialog from '@/features/tasks/components/StrategyFollowUpDialog';
 import { collectSpawnedTemplateIds } from '@/shared/lib/tree-helpers';
