+++ src/features/tasks/components/TaskItem.tsx
@@ -18,7 +18,7 @@ import {
 } from '@/shared/lib/date-engine/formatTaskDueBadge';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/shared/ui/tooltip';
 import { useConfirm } from '@/shared/ui/confirm-dialog';
-import type { PresenceState } from '@/features/projects/hooks/useProjectPresence';
+import type { PresenceState } from '@/shared/types/presence';
 
 export type { TaskItemData } from '@/shared/types/tasks';
 import type { TaskItemData } from '@/shared/types/tasks';
