+++ src/shared/ui/ActivityRow.tsx
